+++ project/src/components/reviews-item/reviews-item.tsx
